+++ Sources/DTTableViewManager/DTTableViewManager.swift
@@ -401,7 +401,6 @@ internal enum EventMethodSignature: String {
     case contextMenuConfigurationForRowAtIndexPath = "tableView:contextMenuConfigurationForRowAtIndexPath:point:"
     case previewForHighlightingContextMenu = "tableView:previewForHighlightingContextMenuWithConfiguration:"
     case previewForDismissingContextMenu = "tableView:previewForDismissingContextMenuWithConfiguration:"
-    case willCommitMenuWithAnimator = "tableView:willCommitMenuWithAnimator:"
     case selectionFollowsFocusForRowAtIndexPath = "tableView:selectionFollowsFocusForRowAtIndexPath:"
     
     /// UITableViewDragDelegate